Add fallback into Linux FB if DRI isn't available
This will automatically try to use the Linux FB driver if it can't
find a suitable DRI device or screen. If it can't create a FB with
double buffering, it will use a single buffer instead.
This will print the escape char to hide the blinking cursor and restore
it at the end, since it would affect the framebuffer. I don't know how
it will react if the terminal doesn't support the escape sequence.

To test on a DRI-enabled machine, simply comment out the line :
//dri = cairo_dri_open("/dev/dri/card0");

Error prints from the linux fb driver still need to be removed, they
are left here for now until further tests are done.